+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -47,7 +47,7 @@ const
 		preserveId: true,
 		restrict: 'self-first'
 	},
-	dataContainerDisabledAttribute = 'data-container-disabled',
+	dataContainerDisabledAttribute = 'data-spotlight-container-disabled',
 	isDown = is('down'),
 	isLeft = is('left'),
 	isRight = is('right'),
@@ -59,9 +59,9 @@ const VirtualListBaseFactory = (type) => {
 	const UiBase = (type === JS) ? UiVirtualListBase : UiVirtualListBaseNative;
 
 	return class VirtualListCore extends Component {
-		static displayName = 'VirtualListBase'
+		/* No displayName here. We set displayName to returned components of this factory function. */
 
-		static propTypes = /** @lends moonstone/VirtualList.VirtualList.prototype */ {
+		static propTypes = /** @lends moonstone/VirtualList.VirtualListBase.prototype */ {
 			/**
 			 * The `render` function for an item of the list receives the following parameters:
 			 * - `data` is for accessing the supplied `data` property of the list.
@@ -92,7 +92,7 @@ const VirtualListBaseFactory = (type) => {
 			 * @type {Function}
 			 * @public
 			 */
-			component: PropTypes.func.isRequired,
+			itemRenderer: PropTypes.func.isRequired,
 
 			/**
 			 * The render function for the items.
@@ -108,7 +108,7 @@ const VirtualListBaseFactory = (type) => {
 			 * @type {String}
 			 * @private
 			 */
-			'data-container-id': PropTypes.string, // eslint-disable-line react/sort-prop-types,
+			'data-spotlight-id': PropTypes.string, // eslint-disable-line react/sort-prop-types,
 
 			/**
 			 * Passes the instance of [VirtualList]{@link ui/VirtualList.VirtualList}.
@@ -130,9 +130,9 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		componentDidMount () {
-			if (type === JS) {
-				const containerNode = this.uiRef.containerRef;
+			const containerNode = this.uiRef.containerRef;
 
+			if (type === JS) {
 				// prevent native scrolling by Spotlight
 				this.preventScroll = () => {
 					containerNode.scrollTop = 0;
@@ -141,14 +141,11 @@ const VirtualListBaseFactory = (type) => {
 
 				if (containerNode && containerNode.addEventListener) {
 					containerNode.addEventListener('scroll', this.preventScroll);
-					containerNode.addEventListener('keydown', this.onKeyDown);
 				}
-			} else {
-				const contentNode = this.uiRef.contentRef;
+			}
 
-				if (contentNode && contentNode.addEventListener) {
-					contentNode.addEventListener('keydown', this.onKeyDown);
-				}
+			if (containerNode && containerNode.addEventListener) {
+				containerNode.addEventListener('keydown', this.onKeyDown);
 			}
 		}
 
@@ -157,20 +154,18 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		componentWillUnmount () {
-			if (type === JS) {
-				const containerNode = this.uiRef.containerRef;
+			const containerNode = this.uiRef.containerRef;
 
+			if (type === JS) {
 				// remove a function for preventing native scrolling by Spotlight
 				if (containerNode && containerNode.removeEventListener) {
 					containerNode.removeEventListener('scroll', this.preventScroll);
 					containerNode.removeEventListener('keydown', this.onKeyDown);
 				}
-			} else {
-				const contentNode = this.uiRef.contentRef;
+			}
 
-				if (contentNode && contentNode.removeEventListener) {
-					contentNode.removeEventListener('keydown', this.onKeyDown);
-				}
+			if (containerNode && containerNode.removeEventListener) {
+				containerNode.removeEventListener('keydown', this.onKeyDown);
 			}
 
 			this.setContainerDisabled(false);
@@ -184,10 +179,10 @@ const VirtualListBaseFactory = (type) => {
 		restoreLastFocused = false
 
 		setContainerDisabled = (bool) => {
-			const containerNode = (type === JS) ? this.uiRef.containerRef : this.uiRef.contentRef;
+			const contentNode = this.uiRef.contentRef;
 
-			if (containerNode) {
-				containerNode.setAttribute(dataContainerDisabledAttribute, bool);
+			if (contentNode) {
+				contentNode.setAttribute(dataContainerDisabledAttribute, bool);
 
 				if (bool) {
 					document.addEventListener('keydown', this.handleGlobalKeyDown, {capture: true});
@@ -330,12 +325,6 @@ const VirtualListBaseFactory = (type) => {
 					isRtl = this.props.rtl,
 					isForward = (direction === 'down' || isRtl && direction === 'left' || !isRtl && direction === 'right');
 
-				if (type === JS) {
-					// To prevent item positioning issue, make all items to be rendered.
-					this.uiRef.updateFrom = null;
-					this.uiRef.updateTo = null;
-				}
-
 				if (firstVisibleIndex <= indexToScroll && indexToScroll <= lastVisibleIndex) {
 					const node = this.uiRef.containerRef.querySelector(`[data-index='${indexToScroll}'].spottable`);
 
@@ -361,7 +350,7 @@ const VirtualListBaseFactory = (type) => {
 		 */
 
 		setRestrict = (bool) => {
-			Spotlight.set(this.props['data-container-id'], {restrict: (bool) ? 'self-only' : 'self-first'});
+			Spotlight.set(this.props['data-spotlight-id'], {restrict: (bool) ? 'self-only' : 'self-first'});
 		}
 
 		setSpotlightContainerRestrict = (keyCode, target) => {
@@ -479,9 +468,7 @@ const VirtualListBaseFactory = (type) => {
 
 			this.isScrolledBy5way = false;
 			if (getDirection(keyCode)) {
-				if (type === Native) {
-					ev.preventDefault();
-				}
+				ev.preventDefault();
 				this.setSpotlightContainerRestrict(keyCode, target);
 				this.isScrolledBy5way = this.jumpToSpottableItem(keyCode, target);
 			}
@@ -511,9 +498,6 @@ const VirtualListBaseFactory = (type) => {
 
 			if (Spotlight.isPaused()) {
 				Spotlight.resume();
-				if (type === JS) {
-					this.forceUpdate();
-				}
 			}
 			this.focusOnNode(item);
 			this.nodeIndexToBeFocused = null;
@@ -523,7 +507,7 @@ const VirtualListBaseFactory = (type) => {
 			if (ref) {
 				if (type === JS) {
 					this.focusOnItem(index);
-				} else if (type === Native) {
+				} else {
 					// If focusing the item of VirtuallistNative, `onFocus` in Scrollable will be called.
 					// Then VirtualListNative tries to scroll again differently from VirtualList.
 					// So we would like to skip `focus` handling when focusing the item as a workaround.
@@ -580,9 +564,9 @@ const VirtualListBaseFactory = (type) => {
 				!this.isPlaceholderFocused()
 			) {
 				const
-					containerId = this.props['data-container-id'],
+					containerId = this.props['data-spotlight-id'],
 					node = this.uiRef.containerRef.querySelector(
-						`[data-container-id="${containerId}"] [data-index="${this.preservedIndex}"]`
+						`[data-spotlight-id="${containerId}"] [data-index="${this.preservedIndex}"]`
 					);
 
 				if (node) {
@@ -625,9 +609,7 @@ const VirtualListBaseFactory = (type) => {
 						node.blur();
 					}
 				}
-				if (type === JS) {
-					this.nodeIndexToBeFocused = null;
-				}
+				this.nodeIndexToBeFocused = null;
 				this.lastFocusedIndex = focusedIndex;
 
 				if (primary.clientSize >= primary.itemSize) {
@@ -692,7 +674,7 @@ const VirtualListBaseFactory = (type) => {
 
 		render () {
 			const
-				{component, itemsRenderer, ...rest} = this.props,
+				{itemRenderer, itemsRenderer, ...rest} = this.props,
 				needsScrollingPlaceholder = this.isNeededScrollingPlaceholder();
 
 			delete rest.initUiChildRef;
@@ -700,14 +682,14 @@ const VirtualListBaseFactory = (type) => {
 			return (
 				<UiBase
 					{...rest}
-					component={({index, ...itemRest}) => ( // eslint-disable-line react/jsx-no-bind
-						component({
+					getComponentProps={this.getComponentProps}
+					itemRenderer={({index, ...itemRest}) => ( // eslint-disable-line react/jsx-no-bind
+						itemRenderer({
 							... itemRest,
 							[dataIndexAttribute]: index,
 							index
 						})
 					)}
-					getComponentProps={this.getComponentProps}
 					ref={this.initUiRef}
 					updateStatesAndBounds={this.updateStatesAndBounds}
 					itemsRenderer={(props) => { // eslint-disable-line react/jsx-no-bind
@@ -734,6 +716,7 @@ const VirtualListBaseFactory = (type) => {
  * @private
  */
 const VirtualListBase = VirtualListBaseFactory(JS);
+VirtualListBase.displayName = 'VirtualListBase';
 
 /**
  * A Moonstone-styled base component for [VirtualListNative]{@link moonstone/VirtualList.VirtualListNative} and
@@ -746,6 +729,7 @@ const VirtualListBase = VirtualListBaseFactory(JS);
  * @private
  */
 const VirtualListBaseNative = VirtualListBaseFactory(Native);
+VirtualListBaseNative.displayName = 'VirtualListBaseNative';
 
 const ScrollableVirtualList = ({role, ...rest}) => ( // eslint-disable-line react/jsx-no-bind
 	<Scrollable